+++ samples/picos/src/3d/sample_3d.hpp
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <sample_base.hpp>
+
+namespace ek {
+
+class Sample3D : public SampleBase {
+public:
+    Sample3D();
+    ~Sample3D() override;
+    void update(float dt) override;
+    void draw() override;
+};
+
+}
+