+++ samples/picos/src/piko/sample_piko.hpp
@@ -1,0 +1,19 @@
+#pragma once
+
+#include <sample_base.hpp>
+
+namespace ek {
+
+class SamplePiko : public SampleBase {
+public:
+    SamplePiko();
+    ~SamplePiko() override;
+
+    void draw() override;
+    void update(float dt) override;
+
+    ecs::entity container{};
+};
+
+}
+