+++ samples/picos/src/sample_base.cpp
@@ -1,0 +1,17 @@
+#include "sample_base.hpp"
+
+namespace ek {
+
+SampleBase::SampleBase() = default;
+
+SampleBase::~SampleBase() = default;
+
+void SampleBase::draw() {
+
+}
+
+void SampleBase::update(float dt) {
+
+}
+
+}