+++ samples/picos/src/sample_base.hpp
@@ -1,0 +1,21 @@
+#pragma once
+
+
+namespace ek {
+
+class SampleBase {
+public:
+    SampleBase();
+
+    virtual ~SampleBase();
+
+    virtual void draw();
+
+    virtual void update(float dt);
+
+private:
+
+};
+
+}
+